+++ modules/hanna-css/src/lib/style-server-info.ts
@@ -69,6 +69,7 @@ export type CssModuleToken =
   | 'MainMenu'
   | 'MiniMetrics'
   | 'Modal'
+  | 'Multiselect'
   | 'NameCard'
   | 'NameCards'
   | 'NewsHero'