+++ modules/hanna-react/CHANGELOG.md
@@ -3,10 +3,16 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. -->
+- feat: Add component `Multiselect`
 - `Datepicker`:
   - feat: Add support for "uncontrolled" mode — (add prop `defaultValue`, make
     `value` and `onChange` optional.)
   - feat: Add prop `isoMode` to generate/submit ISO-8601 `<input/>` values
+- Checkboxes and Radio buttons
+  - feat: Support `readOnly` — using `disabled` + `input[type=hidden]`
+  - feat: Support passing `options` as simple string array
+  - feat: Make `name` prop optional for groups
+  - fix: Handling of `disabled` as array of indexes
 
 ## 0.10.92